Time: 0 ms (100.00%), Space: 10.8 MB (55.43%) - LeetHub

+++ 0129-sum-root-to-leaf-numbers/0129-sum-root-to-leaf-numbers.cpp
@@ -11,7 +11,7 @@
  */
 class Solution {
 public:
-    long long ans = 0;
+    int ans = 0;
     string str;
     
     void dfs(TreeNode* root) {
@@ -26,6 +26,7 @@ public:
         
         dfs(root->left);
         dfs(root->right);
+        
         str.pop_back();
         
     }
